add error_reporting.py
