+++ openadapt/config.py
@@ -8,17 +8,13 @@ import multiprocessing
 import os
 import pathlib
 import shutil
-import webbrowser
 
 from loguru import logger
 from pydantic import field_validator
 from pydantic.fields import FieldInfo
 from pydantic_settings import BaseSettings, PydanticBaseSettingsSource
-from PySide6.QtWidgets import QMessageBox, QPushButton
-import git
-import sentry_sdk
 
-from openadapt.build_utils import get_root_dir_path, is_running_from_executable
+from openadapt.build_utils import get_root_dir_path
 
 CONFIG_DEFAULTS_FILE_PATH = (
     pathlib.Path(__file__).parent / "config.defaults.json"
@@ -411,47 +407,3 @@ def print_config() -> None:
             if not key.startswith("_") and key.isupper():
                 val = maybe_obfuscate(key, val)
                 logger.info(f"{key}={val}")
-
-        if config.ERROR_REPORTING_ENABLED:
-            if is_running_from_executable():
-                is_reporting_branch = True
-            else:
-                active_branch_name = git.Repo(PARENT_DIR_PATH).active_branch.name
-                logger.info(f"{active_branch_name=}")
-                is_reporting_branch = (
-                    active_branch_name == config.ERROR_REPORTING_BRANCH
-                )
-                logger.info(f"{is_reporting_branch=}")
-            if is_reporting_branch:
-
-                def show_alert() -> None:
-                    """Show an alert to the user."""
-                    msg = QMessageBox()
-                    msg.setIcon(QMessageBox.Warning)
-                    msg.setText("""
-                        An error has occurred. The development team has been notified.
-                        Please join the discord server to get help or send an email to
-                        help@openadapt.ai
-                        """)
-                    discord_button = QPushButton("Join the discord server")
-                    discord_button.clicked.connect(
-                        lambda: webbrowser.open("https://discord.gg/yF527cQbDG")
-                    )
-                    msg.addButton(discord_button, QMessageBox.ActionRole)
-                    msg.addButton(QMessageBox.Ok)
-                    msg.exec()
-
-                def before_send_event(event: Any, hint: Any) -> Any:
-                    """Handle the event before sending it to Sentry."""
-                    try:
-                        show_alert()
-                    except Exception:
-                        pass
-                    return event
-
-                sentry_sdk.init(
-                    dsn=config.ERROR_REPORTING_DSN,
-                    traces_sample_rate=1.0,
-                    before_send=before_send_event,
-                    ignore_errors=[KeyboardInterrupt],
-                )
