+++ openadapt/entrypoint.py
@@ -9,6 +9,7 @@ if __name__ == "__main__":
 from loguru import logger
 
 from openadapt.build_utils import redirect_stdout_stderr
+from openadapt.error_reporting import configure_error_reporting
 
 
 def run_openadapt() -> None:
@@ -20,6 +21,7 @@ def run_openadapt() -> None:
             from openadapt.config import print_config
 
             print_config()
+            configure_error_reporting()
             load_alembic_context()
             tray._run()
         except Exception as exc:
